in2/requirements.txt
aws-cdk.core
aws-cdk.aws_ecr
